fix: satisfy mypy no-any-return in _realized_vol_pct
Pre-existing pstdev(...) return type ambiguity, unrelated to the
market-regime fix but now blocking CI's changed-file mypy gate since
this file is touched by this PR. Explicit float() cast, no behavior
change.

+++ apps/api/app/services/market_regime.py
@@ -230,7 +230,7 @@ class MarketRegimeService:
             returns.append(float((curr - prev) / prev))
         if len(returns) < 2:
             return 0.0
-        return pstdev(returns) * 100 * (252**0.5)
+        return float(pstdev(returns) * 100 * (252**0.5))
 
     def _breadth_pct(self, snapshots: list[_SeriesSnapshot]) -> float:
         votes = 0
